requirements.txt
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
fastapi==0.115.12
uvicorn==0.34.0
numpy==2.2.4
bcrypt==4.3.0
pytest==8.3.5
httpx==0.28.1
mlflow==2.21.3
streamlit==1.44.1
requests==2.32.3
azureml_core==1.60.0
azureml-dataprep==5.3.3
azure-storage-blob==12.25.1
azure-identity==1.15.0